Enable renaming a selection option with custom labels
