+++ web_custom_label/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     'name': 'Web Custom Label',
-    'version': '1.0.0',
+    'version': '1.1.0',
     'author': 'Numigi',
     'maintainer': 'Numigi',
     'website': 'https://bit.ly/numigi-com',
