+++ web_custom_label/i18n/fr.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 11.0+e\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-10-18 21:07+0000\n"
-"PO-Revision-Date: 2018-10-18 17:12-0400\n"
+"POT-Creation-Date: 2020-02-06 19:43+0000\n"
+"PO-Revision-Date: 2020-02-06 14:43-0500\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "Language: fr\n"
@@ -20,7 +20,7 @@ msgstr ""
 #. module: web_custom_label
 #: model:ir.model.fields,field_description:web_custom_label.field_web_custom_label__active
 msgid "Active"
-msgstr ""
+msgstr "Actif"
 
 #. module: web_custom_label
 #: model:ir.model,name:web_custom_label.model_base
@@ -73,6 +73,11 @@ msgid "ID"
 msgstr "ID"
 
 #. module: web_custom_label
+#: model:ir.model.fields,field_description:web_custom_label.field_web_custom_label__key
+msgid "Key"
+msgstr "Clé"
+
+#. module: web_custom_label
 #: selection:web.custom.label,position:0
 msgid "Label"
 msgstr "Libellé"
@@ -119,6 +124,11 @@ msgid "Reference"
 msgstr "Référence"
 
 #. module: web_custom_label
+#: selection:web.custom.label,position:0
+msgid "Selection"
+msgstr "Sélection"
+
+#. module: web_custom_label
 #: model:ir.model.fields,field_description:web_custom_label.field_web_custom_label__term
 msgid "Term"
 msgstr "Terme"
@@ -136,7 +146,7 @@ msgstr "Type"
 #. module: web_custom_label
 #: model:ir.model,name:web_custom_label.model_ir_ui_view
 msgid "View"
-msgstr ""
+msgstr "Vue"
 
 #. module: web_custom_label
 #: selection:web.custom.label,type_:0
