+++ web_custom_label/models/__init__.py
@@ -1,0 +1,8 @@
+# © 2018 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from . import (
+    base,
+    ir_ui_view,
+    web_custom_label,
+)
